+++ crates/objc2/src/top_level_traits.rs
@@ -369,7 +369,7 @@ pub trait DefinedClass: ClassType {
 ///     }
 ///
 ///     impl MyClass {
-///         #[unsafe(method_id(init))]
+///         #[unsafe(method(init))]
 ///         fn init(this: Allocated<Self>) -> Retained<Self> {
 ///             let this = this.set_ivars(Ivars::<Self> {
 ///                 // Initialize to some default value.