+++ crates/test-ui/ui/wrong_optional.rs
@@ -42,7 +42,7 @@ define_class!(
     impl CustomObject2 {
         #[optional]
         /// Doc comment
-        #[unsafe(method_id(d))]
+        #[unsafe(method(d))]
         fn d(&self) -> Retained<Self> {
             unimplemented!()
         }